% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wa_get_registers_by_id}
\alias{wa_get_registers_by_id}
\title{Fetch all pages from \code{getRegistrosJSON} results.}
\usage{
wa_get_registers_by_id(id, tm = c("s", "f"), sys_sleep = 0.01)
}
\arguments{
\item{id}{integer. Wikiaves' id of a species.}

\item{tm}{sound ("s") or photo ("f")}

\item{sys_sleep}{numeric. A time offset to avoid exploding wikiaves' server}
}
\description{
Fetch all pages from \code{getRegistrosJSON} results.
}
