% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wa_get_registros_json}
\alias{wa_get_registros_json}
\title{Interface to \code{getRegistrosJSON} from wikiaves}
\usage{
wa_get_registros_json(id, p = 1, tm = c("s", "f"), t = "s")
}
\arguments{
\item{id}{integer. Wikiaves' id of a species.}

\item{p}{integer. The page number to fetch the results.}

\item{tm}{sound ("s") or photo ("f").}

\item{t}{unkown yet.}
}
\description{
Interface to \code{getRegistrosJSON} from wikiaves
}
