% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wa_get_registros_json_n}
\alias{wa_get_registros_json_n}
\title{Title}
\usage{
wa_get_registros_json_n(id, tm = c("s", "f"))
}
\arguments{
\item{id}{integer. Wikiaves' id of a species.}

\item{tm}{sound ("s") or photo ("f")}
}
\description{
Title
}
