% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wa_metadata}
\alias{wa_metadata}
\title{Main function. Get metadata from wikiaves from a set of terms.}
\usage{
wa_metadata(term, tm = c("s", "f"), verbose = TRUE, download = FALSE,
  path = getwd(), mp3_file_name = "{label}-{id1}.mp3", parallel = 1,
  metadata_sys_sleep = 0.1, download_sys_sleep = 1e-04,
  force = FALSE)
}
\arguments{
\item{term}{character vector with the terms to be searched by wikiaves API. Can be a full species specification or just part of it, eg. c("megascops choliba", "strix h").}

\item{tm}{sound ("s") or photo ("f")}

\item{verbose}{boolean. Should it give extra information during the process? Default to TRUE.}

\item{download}{boolean. Should the files be downloaded? Default to FALSE. You can choose to download later by using wa_download() function.}

\item{path}{character. A string specifying the place to store the MP3 files if download is set to TRUE.}

\item{mp3_file_name}{character. A 'glue'-like format string with the pattern of the names of the mp3 files.}

\item{parallel}{integer. Number of cores for parallel processing.}

\item{metadata_sys_sleep}{numeric. A time offset to avoid exploding wikiaves' server.}

\item{download_sys_sleep}{numeric. A time offset to avoid exploding wikiaves' server.}

\item{force}{boolean. If mp3 exists, should it be overwritten? If set to FALSE it will download non-existing or zero sized files only. Default to FALSE.}
}
\description{
Main function. Get metadata from wikiaves from a set of terms.
}
